+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version: 1.66
+
+- [feature] Cache Venus::Meta for better performance
+- [update] Update Venus::Role::Optional
+
 ## Version: 1.65
 
 - [feature] Implement Venus::Gather